+++ src/components/MartiniContainer/MartiniContainer.tsx
@@ -1,7 +1,7 @@
 import {FC} from "react"
 import "./MartiniContainer.css"
 import MartiniImage from "../../../src/Martini.jpeg"
-import MartiniCard from "../MartiniCard/MartiniCard"
+import MartiniDeck from "../MartiniDeck/MartiniDeck"
 
 
 type SpecialtyState = {
@@ -32,30 +32,32 @@ const MartiniContainer: FC<SpecialtyState> = ({ martiniDrinks }) => {
 
    let allMartinis =  martiniDrinks.map((martini: SpecialtyDrink) => {
             return (
-            <MartiniCard 
-                    idDrink={martini.idDrink}   
+            <MartiniDeck
+                idDrink={martini.idDrink}   
                 strDrink={martini.strDrink}
                 strDrinkThumb={martini.strDrinkThumb}
-                strInstructions={martini.strInstructions}
-                strIngredient1={martini.strIngredient1}
-                strIngredient2={martini.strIngredient2}
-                strIngredient3={martini.strIngredient3}
-                strIngredient4={martini.strIngredient4}
-                strIngredient5={martini.strIngredient5}
-                strIngredient6={martini.strIngredient6}
-                strIngredient7={martini.strIngredient7}
-                strIngredient8={martini.strIngredient8}
-                strIngredient9={martini.strIngredient9}
-                strIngredient10={martini.strIngredient10}
-                strIngredient11={martini.strIngredient11}
+                // strInstructions={martini.strInstructions}
+                // strIngredient1={martini.strIngredient1}
+                // strIngredient2={martini.strIngredient2}
+                // strIngredient3={martini.strIngredient3}
+                // strIngredient4={martini.strIngredient4}
+                // strIngredient5={martini.strIngredient5}
+                // strIngredient6={martini.strIngredient6}
+                // strIngredient7={martini.strIngredient7}
+                // strIngredient8={martini.strIngredient8}
+                // strIngredient9={martini.strIngredient9}
+                // strIngredient10={martini.strIngredient10}
+                // strIngredient11={martini.strIngredient11}
             />);
             })
         
             return(
-                 <div className="martiniButton">
-                    {allMartinis}
-                    <img className="martiniImage" src={MartiniImage} />
-                    <h3>10 Martinis that will impress your guests</h3>
+                <div>
+                   {/* {allMartinis} */}
+                    <div className="martiniButton">
+                        <img className="martiniImage" src={MartiniImage} />
+                        <h3>10 Martinis that will impress your guests</h3>
+                    </div>
                 </div>
             )
 }
