+++ src/components/MartiniDeck/MartiniDeck.css
@@ -1,0 +1,51 @@
+
+/* img {
+    height: 20%;
+    width: 20%
+} */
+
+/* .details {
+border: 1px solid black;
+display:flex;
+flex-direction: row;
+justify-content: center;
+align-items: center;
+align-content: center;
+width: 50%;
+} */
+
+/* .ingredients {
+border: 2px solid pink;
+width: 20%;
+height: 20%;
+}
+
+.instructions {
+border: 2px solid pink;
+width: 80%;
+height: 20%;
+} */
+
+
+.martiniContainer {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    /* grid-template-rows: repeat(4, 1fr); */
+    grid-column-gap: 20px;
+    grid-row-gap: 20px;
+    margin-left: 120px;
+    margin-top: 60px;
+    margin-bottom: 70px;
+}
+
+.drink-card {
+    border: 2px solid black;
+}
+
+.drink-image {
+    display: flex;
+    flex-direction: column;
+    height: 80%;
+    width: 80%;
+    margin-top: 10%;
+}
